refactor(kikan): hoist Mokumo-vocab admin/auth/diagnostics to mokumo-shop; extend I1 strict [#617 #618]
Completes the Graft capability/vocabulary split. Kikan stops naming
`SetupMode` and `"mokumo.db"` in production code; every Mokumo-vocab
handler module (admin UDS router, diagnostics, profile list, profile
switch, demo reset, auth HTTP, user admin) moves to `mokumo-shop`.

Hoists from `crates/kikan/src/` to `crates/mokumo-shop/src/`:
- `control_plane/diagnostics.rs` + `platform/{diagnostics,diagnostics_bundle}.rs` → `admin/{diagnostics,diagnostics_http,diagnostics_bundle_http}.rs` (DTO names `production`+`demo` + `SetupMode` field + Mokumo-branded zip filename)
- `control_plane/profile_list.rs` → `admin/profile_list.rs` (names SetupMode in the wire DTO)
- `engine::admin` module → `admin/router.rs` (full UDS router — mounts diagnostics + profile list + switch + backup + migrate status)
- `platform/demo.rs` → `demo_reset.rs` (Mokumo-specific demo-reset + sidecar handlers)
- `platform/auth/{mod,reset,recover}.rs` → `auth_handlers/{mod,reset,recover}.rs` (admin@demo.local literal, demo auto-login, SetupMode::Production auth target)
- `platform/users/mod.rs` → `user_admin.rs` (handlers use AuthSession<Backend<SetupMode>> concretion)
- `tenancy/layout.rs` → DELETED (dead duplicate; live copy always in mokumo-shop startup)

Kikan generic-K cleanups:
- `switch_profile_admin<K>` returns `(K, K)`; mokumo-shop wraps into `ProfileSwitchAdminResponse`.
- `Engine::admin_router` + `platform_protected_routes` + `kikan::admin::*` re-export removed — mokumo-server now calls `mokumo_shop::admin::build_admin_router` directly.
- Doc-comment sweep removes remaining `SetupMode` / `"mokumo.db"` prose mentions from kikan/src.
- In-file test modules in `backup/{mod,snapshot,restore}.rs` relocate to sibling `_tests.rs` files via `#[path = "..."]` so the I1-strict file-level glob stays clean.

I1 CI strengthening:
- `scripts/check-i1-domain-purity.sh` gains a strict pattern class: `\bSetupMode\b|"mokumo\.db"` under `crates/kikan/src/` excluding `*_tests.rs`. Classic I1 = shop nouns; strict I1 = leaked wire artifacts. `crates/kikan/CLAUDE.md` documents the split.

Observable wire change: `DiagnosticsResponse.app.name` now reports `"mokumo-shop"` instead of `"kikan"` — `env!("CARGO_PKG_NAME")` resolves to the crate owning the snapshot fn. Route shapes, JSON fields, zip contents, BDD contracts unchanged.

ACs green:
- `rg '\bSetupMode\b' crates/kikan/src/ -g '!*_tests.rs'` → 0 hits
- `rg '"mokumo\.db"' crates/kikan/src/ -g '!*_tests.rs'` → 0 hits
- `bash scripts/check-i*.sh` all green (classic + strict I1, I2, I3, I4, I5)
- `cargo test -p kikan -p mokumo-shop --lib --tests` green (219 + 170 lib + integration suites)
- `cargo deny check licenses bans sources` green (new deps: anyhow, sysinfo, zip, rand)
