+++ crates/kikan/src/control_plane/mod.rs
@@ -21,9 +21,7 @@
 //! one-shot CLI owns tty prompts — none of which leak down into the
 //! pure-fn layer.
 
-pub mod diagnostics;
 pub mod migration_status;
-pub mod profile_list;
 pub mod profiles;
 pub mod state;
 pub mod users;
