+++ crates/mokumo-shop/src/admin/diagnostics_bundle_http.rs
@@ -1,13 +1,12 @@
 //! HTTP delegation for `GET /api/diagnostics/bundle`. The zip-build +
-//! log-redaction logic lives in `kikan::control_plane::diagnostics::
-//! build_bundle`; this module only adds the HTTP response headers.
+//! log-redaction logic lives in [`crate::admin::diagnostics::build_bundle`];
+//! this module only adds the HTTP response headers.
 
 use axum::{extract::State, http::header, response::IntoResponse};
-
-use crate::{AppError, PlatformState, control_plane};
+use kikan::{AppError, PlatformState};
 
 pub async fn handler(State(state): State<PlatformState>) -> Result<impl IntoResponse, AppError> {
-    let (zip_bytes, filename) = control_plane::diagnostics::build_bundle(&state).await?;
+    let (zip_bytes, filename) = crate::admin::diagnostics::build_bundle(&state).await?;
 
     Ok((
         [
