+++ crates/mokumo-shop/tests/auth_reset_regressions.rs
@@ -4,8 +4,8 @@ use axum_test::TestServer;
 use kikan::auth::SeaOrmUserRepo;
 use kikan::auth::UserRepository;
 use kikan::auth::{CreateUser, RoleId};
-use kikan::platform::auth::reset::recovery_file_path_for_email;
 use kikan_types::SetupMode;
+use mokumo_shop::auth_handlers::reset::recovery_file_path_for_email;
 use mokumo_shop::startup::ensure_data_dirs;
 use sea_orm::DatabaseConnection;
 use serde_json::json;
