+++ crates/mokumo-shop/tests/platform_startup_guards.rs
@@ -445,3 +445,44 @@ fn all_migrations_use_transaction_returns_some_true() {
         );
     }
 }
+
+// ─── migrate_flat_layout ─────────────────────────────────────────────────────
+
+#[test]
+fn migrate_flat_layout_noop_when_no_flat_db() {
+    let tmp = tempfile::tempdir().unwrap();
+    mokumo_shop::startup::migrate_flat_layout(tmp.path()).unwrap();
+}
+
+#[test]
+fn migrate_flat_layout_moves_flat_to_production() {
+    let tmp = tempfile::tempdir().unwrap();
+    let flat_db = tmp.path().join("mokumo.db");
+    {
+        let conn = rusqlite::Connection::open(&flat_db).unwrap();
+        conn.execute_batch("CREATE TABLE test (id INTEGER PRIMARY KEY)")
+            .unwrap();
+    }
+    mokumo_shop::startup::migrate_flat_layout(tmp.path()).unwrap();
+
+    assert!(!flat_db.exists(), "flat db should be removed");
+    assert!(
+        tmp.path().join("production/mokumo.db").exists(),
+        "production db should exist"
+    );
+    let profile = std::fs::read_to_string(tmp.path().join("active_profile")).unwrap();
+    assert_eq!(profile, "production");
+}
+
+#[test]
+fn migrate_flat_layout_idempotent() {
+    let tmp = tempfile::tempdir().unwrap();
+    std::fs::create_dir_all(tmp.path().join("production")).unwrap();
+    {
+        let conn = rusqlite::Connection::open(tmp.path().join("production/mokumo.db")).unwrap();
+        conn.execute_batch("CREATE TABLE test (id INTEGER PRIMARY KEY)")
+            .unwrap();
+    }
+    std::fs::write(tmp.path().join("active_profile"), "production").unwrap();
+    mokumo_shop::startup::migrate_flat_layout(tmp.path()).unwrap();
+}
